Patents

Patent documents for Covid-19 related activity from the Lens database.

patents

Format

A data table with 40 variables and 43,075 rows.

kind

Patent kind code https://worldwide.espacenet.com/help?locale=en_EP&method=handleHelpTopic&topic=kindcodes

lens_publication_number

publication number in Lens format with spaces and "\"

lens_id

Lens database unique identifier, use to upload selections to the Lens

publication_date

patent publication data as YYYY-MM-DD

publication_year

patent publication year

lens_application_number

patent application number in Lens format with spaces and "\"

application_date

patent application date as YYYY-MM-DD

priority_numbers

patent priority numbers, concatenated with ";;"

earliest_priority_date

Earliest priority date for a publication

title

Patent title

applicants

Patent Applicants, concatenated with ";;"

inventors

Inventors, concatenated with ";;"

owners_us

Patent owners for the US, may differ from applicants and may be incomplete

url

URL to the Lens record

type

Document Type, may be an application, grant, search report, design and other categories

has_full_text

Whether the text is available from the lens

cited_by_patent_count

How many times a document is cited by a later patent document

simple_family_size

Number of simple patent family

extended_family_size

Number of extended patent family members

sequence_count

Number of DNA/amino acid sequences accompanying the document

cpc_classifications

Cooperative Patent Classification codes

ipcr_classifications

International Patent Classification codes

us_classifications

defunkt US classification codes, replaced by the CPC

npl_citation_count

count of non-patent literature citations in the document

npl_resolved_citation_count

Number of non-patent literature citations that can be resolved

npl_resolved_lens_id

Lens identifier for resolved non-patent literature citations

npl_resolved_external_id

number of resolved external non-patent literature citations

npl_citations

raw non-patent literature citations

source

The Lens Covid-19 source file name

lens_jurisdiction

The jurisdiction where a document was published, the publication country

publication_country

the country of publication

publication_number

the publication number, spaces and "\" removed for cross database use

application_number

the patent application number, may contain duplicates for publications

application_country

the country of application

application_year

the application year

publication_country_name

the full name of the country/regional/international patent office

application_country_name

the full name of the country/regional/international patent office

dataset

the short name for the source dataset listed in source

data_type

patent, literature or sequence

Source

Lens Covid-19 Patent datasets. Created by Osmat Jefferson. For further details see https://about.lens.org/covid-19/